+++ src/containers/treePage/ducks/selectors.ts
@@ -48,6 +48,7 @@ export const mapStewardshipToTreeCare = (
         month,
         year,
         message: `Was ${activityStrings.join(' and ')}.`,
+        id: item.id,
       };
     });
   }
