+++ src/containers/treePage/test/selectors.test.ts
@@ -60,6 +60,7 @@ describe('Tree Page Selectors', () => {
           month: 'Jan',
           year: 2021,
           message: 'Was mulched and watered and weeded.',
+          id: 0,
         },
         {
           activityId: 1,
@@ -68,6 +69,7 @@ describe('Tree Page Selectors', () => {
           month: 'Feb',
           year: 2021,
           message: 'Was cleared of waste.',
+          id: 1,
         },
       ];
